+++ app/controllers/api/v1/qcs_controller.rb
@@ -8,11 +8,16 @@ module Api
       def index
         authorize current_player
         @pagy, @qcs = pagy(Qc.includes(%i[qc_attacks qc_charms qc_merits poisons spells]).where(player_id: current_player.id))
+
+        return unless stale? @qcs
+
         render json: @qcs
       end
 
       def show
         authorize @qc
+
+        return unless stale? @qc
 
         if policy(@qc).update?
           render json: @qc