+++ app/javascript/lca/components/battlegroups/battlegroupCreatePopup.jsx
@@ -58,6 +58,11 @@ class BattlegroupCreatePopup extends React.Component<Props, State> {
               margin="normal"
               fullWidth
               onChange={handleChange}
+              inputProps={{
+                autocomplete: 'off',
+                'data-1p-ignore': 'true',
+                'data-lp-ignore': 'true',
+              }}
             />
           </DialogContent>
           <DialogActions>
@@ -77,9 +82,6 @@ class BattlegroupCreatePopup extends React.Component<Props, State> {
   }
 }
 
-const enhance: Enhancer<Props, {}> = connect(
-  null,
-  { createBattlegroup }
-)
+const enhance: Enhancer<Props, {}> = connect(null, { createBattlegroup })
 
 export default enhance(BattlegroupCreatePopup)