+++ spec/models/lunar_character_spec.rb
@@ -35,16 +35,5 @@ RSpec.describe LunarCharacter do
   #   end
   # end
 
-  describe 'converting types' do
-    %i[
-      character solar_character dragonblood_character sidereal_character
-      custom_ability_character custom_attribute_character custom_essence_character
-    ].each do |char|
-      it "works for #{char}" do
-        g = create(char)
-        m = described_class.from_character!(g)
-        expect(m).to be_valid
-      end
-    end
-  end
+  include_examples 'convertable_character', :lunar_character
 end